+++ master-service/src/main/java/me/catzy/invester/master/controller/ArticleController.java
@@ -53,7 +53,7 @@ public class ArticleController extends GenericController<ArticleEntity, Long> {
 	}
 	
 	@GetMapping("chid")
-	public int getChId() {
+	public long getChId() {
 		return service.getChid();
 	}
 }
